migrate real time conditions
