+++ src/main/java/org/betonquest/betonquest/quest/condition/realtime/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * {@link org.betonquest.betonquest.api.quest.condition.Condition Condition} implementations of real time conditions.
+ */
+package org.betonquest.betonquest.quest.condition.realtime;
